requirements.txt
dlib
imutils
mtcnn
numpy
omegaconf
opencv-python
Pillow
scikit-learn
tensorflow-gpu
torch==1.11.0
torchvision==0.12.0
tqdm